+++ packages/x-components/src/x-modules/popular-searches/store/__tests__/getters.spec.ts
@@ -7,6 +7,7 @@ import { getSuggestionsStub } from '../../../../__stubs__/suggestions-stubs.fact
 import { getMockedAdapter, installNewXPlugin } from '../../../../__tests__/utils';
 import { popularSearchesXStoreModule } from '../module';
 import { PopularSearchesState } from '../types';
+import { SuggestionsRequest } from '../../../../../../search-adapter/src/types/requests.types';
 import { resetPopularSearchesStateWith } from './utils';
 
 describe('testing popular searches module getters', () => {
@@ -16,8 +17,16 @@ describe('testing popular searches module getters', () => {
 
   describe(`${gettersKeys.request} getter`, () => {
     it('should return a request object with config default values', () => {
-      resetPopularSearchesStateWith(store, { config: { maxItemsToRequest: 3 } });
-      expect(store.getters[gettersKeys.request]).toEqual({ rows: 3, start: 0 });
+      resetPopularSearchesStateWith(store, {
+        config: { maxItemsToRequest: 3 },
+        params: {
+          catalog: 'es'
+        }
+      });
+      expect(store.getters[gettersKeys.request]).toEqual<
+        SuggestionsRequest & { [key: string]: unknown }
+        // TODO - Remove when the facets refactor is completed.
+      >({ rows: 3, start: 0, catalog: 'es' });
     });
   });